// src/app/model/dependentdataconfig.ts
export class Dependentdataconfig {
    slNo: number;
    type: string;
    parentAtt:string;
    parentValue: string;
    childAtt:string;
    childDataType: string;
    childValue: string;
    childDataSource: string;
    parentType: string;
    workLocation: string;
    orgId: number;
    id: number;
    workLocationId: number;
    bp: number;
    level: string;
    task: number;
}